server/requirements.txt
amqp==5.0.6
asgiref==3.4.1
autopep8==1.6.0
aws-adfs==1.13.0
billiard==3.6.4.0
boto3==1.19.9
botocore==1.22.9
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.28.0
configparser==5.0.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
django-environ==0.8.1
django-redis==5.0.0
djangorestframework==3.12.4
future==0.18.2
idna==3.3
jmespath==0.10.0
kombu==5.2.2
logging518==0.2.2
lxml==4.6.4
packaging==21.2
pdf2image==1.16.0
Pillow==8.4.0
prompt-toolkit==3.0.23
pycodestyle==2.8.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
redis==4.0.2
requests==2.26.0
s3transfer==0.5.0
setuptools-scm==6.3.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
